+++ src/pages/Agents.tsx
@@ -1,5 +1,4 @@
 import useAuthStore from "@/features/auth/useAuthStore";
-import Section from "@/layouts/Section";
 import { pb } from "@/lib/db/pocketbase";
 import { Redirect } from "wouter";
 
@@ -9,11 +8,5 @@ export default function Agents() {
   const isAdmin = pb.authStore?.isSuperuser && isAuth;
   if (!isAdmin) return <Redirect to="/" />;
 
-  return (
-    <Section>
-      <div className="p-2">
-        <h1>Agents</h1>
-      </div>
-    </Section>
-  );
+  return <h1>Agents</h1>;
 }